Fix required engine version in package.json
Currently, on 5.5.0 for example, you can install the package without a warning but running the code produces a SyntaxError (see [Issue #5](https://github.com/Unibeautify/unibeautify-cli/issues/5)). Use of new language features require a more recent engine version as compared to the one specified as a min. requirement in the package.json. This change corrects the engine version requirement in order to warn users with incompatible versions during installation.